backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Kethan Vegunta <kethan@vegunta.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-cors = "^4.0.1"
python-dotenv = "^1.0.1"
supertokens-python = "^0.22.1"
flask-pymongo = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
